Add files via upload
1A.
Aim:
Write a program to accept ‘n’ numbers into an array and then calculate the sum of numbers present in odd positions and even positions respectively.
____
1B.
Aim:
Write a program to implement a Binary Search algorithm.
Write a search function which takes a SearchList as its first parameter and a Comparable as its second.
If either parameter is null, or if the SearchList is empty, you should return NULL.

Implement the following algorithm:

Examine the value in the middle of the current array and print it.

If the midpoint value is the value that we are looking for, return true.

If the value that we are looking for is greater than the midpoint value, adjust the current array to start at the midpoint and print the index.

If the value that we are looking for is less than the midpoint value, adjust the current array to end at the midpoint and print the index.

Continue until you find the value, or until the start reaches the end.
_____
1C.
Aim:
Write a program which creates a structure Student that must have the attribute Avg Marks with 3 more attributes.
Sort the list of the students in descending order using Bubble Sort.
_____
2.
Aim:
Design, Develop and Implement a menu driven Program in C for the following operations on STACK of Integers
(Array Implementation of Stack with maximum size MAX):

a. Push an Element onto Stack
b. Pop an Element from Stack
c. Demonstrate how Stack can be used to check Palindrome
d. Demonstrate Overflow and Underflow situations on Stack
e. Display the status of Stack
f. Exit

Support the program with appropriate functions for each of the above operations.
______

3.
Aim:
Design, Develop and Implement a menu driven Program in C for the following operations on QUEUE of Characters
(Array Implementation of Queue with maximum size MAX):

a. Insert an Element onto QUEUE
b. Delete an Element from QUEUE
c. Demonstrate Overflow and Underflow situations on QUEUE
d. Display the status of QUEUE
e. Exit

Support the program with appropriate functions for each of the above operations.
_____
4.
Aim:
This C program creates a linked list to store integer elements. It prompts the
user to enter elements and add them to the list until the user enters 0. It then
traverses the list and prints each element and "=>" until reaching the null
pointer. Finally, it displays the number of nodes in the list.
______
5.
Aim:
Write a program to insert Number of Nodes in the Binary Tree and Traverse in lnorder
Preorder and Post Order and Search an element in Binary Tree(Display NULL if nor
found, If found Display Found)
_______
6.
Aim:
Implement Binary search tree(BST) with following Menu operations
1.
earch an element in BST(Display NULL if not found, If found Display Found)
2. I nsert an element in BST
3.
Delete leaf element in BST
4.Exit
_____
7.
Aim:
Consider the undirected graph G, consisting of n nodes laid out in a 3 -by- 3 grid:
Start searching at node 1, and break ties for exploring the next node based on lower
numerical order (i.e. add nodes to a queue low to high, add nodes to a stack high to
low). (a) In what order are nodes marked as explored by BFS? (b) In what order are
nodes marked as explored by DFS?
______
8.
Aim:
Write a program to implement Kruskal's Algorithm inorder to find the Minimum Spanning Tree(MST) of a connected,weighted and undirected graph.
______
9.
Aim:
Assume that you have a seven-slot closed hash table (the slots are numbered o
through 6) Write program to final.
________
10.
Aim:
Write a C program (name it "filecopy. c", that copies the contents of one file
to a destination file. This program will read data from one file and copy them
to another, The first input that the program will need is the names of the two
files: input file ("input.txt") and output file ("output.txt"). Once the two file
names have been obtained, the program must open the input fle and create
and open the output file.